Decay models added to fit_decay function
All decays models are now added to the fit_decay function in pydamm.
Notice that default parameters might be a very bad guess and for a
second in chain nucleus the fit might converge to a strange values
with a bad starting point. In such a case an old method of creating
list of parameters (with better guesses) is recommended.